Extract lock functionality into module and use it to prevent race conditions in UploadSets. Fixes #560
