+++ curation_concerns-models/app/models/concerns/curation_concerns/upload_set_behavior.rb
@@ -12,26 +12,18 @@ module CurationConcerns
     end
 
     module ClassMethods
+      include CurationConcerns::Lockable
+
+      # This method handles most race conditions gracefully.
       def find_or_create(id)
-        UploadSet.find(id)
-      rescue ActiveFedora::ObjectNotFoundError
-        safe_create(id)
-      end
-
-      private
-
-        # This method handles most race conditions gracefully.
-        # If an upload_set with the same ID is created by another thread
-        # we fetch the upload_set that was created (rather than throwing
-        # an error) and continute.
-        def safe_create(id)
-          UploadSet.create(id: id)
-        rescue ActiveFedora::IllegalOperation
-          # This is the exception thrown by LDP when we attempt to
-          # create a duplicate object. If we can find the object
-          # then we are good to go.
-          UploadSet.find(id)
+        acquire_lock_for(id) do
+          begin
+            UploadSet.find(id)
+          rescue ActiveFedora::ObjectNotFoundError
+            UploadSet.create(id: id)
+          end
         end
+      end
     end
   end
 end
