+++ curation_concerns-models/app/services/curation_concerns/lock_manager.rb
@@ -15,13 +15,15 @@ module CurationConcerns
 
     # Blocks until lock is acquired or timeout.
     def lock(key)
+      returned_from_block = nil
       client.lock(key, @ttl) do |locked|
         if locked
-          yield
+          returned_from_block = yield
         else
           raise UnableToAcquireLockError
         end
       end
+      returned_from_block
     end
 
     private
