+++ curation_concerns-models/lib/curation_concerns/models/engine.rb
@@ -11,6 +11,7 @@ module CurationConcerns
         require 'active_fedora/noid'
         require 'curation_concerns/noid'
         require 'curation_concerns/permissions'
+        require 'curation_concerns/lockable'
       end
 
       initializer 'configure' do
